fix: move ctre config file
