+++ src/main/java/frc/lib/CTREConfigs.java
@@ -1,10 +1,11 @@
-package frc.robot;
+package frc.lib;
 
 import com.ctre.phoenix6.configs.CANcoderConfiguration;
 import com.ctre.phoenix6.configs.MagnetSensorConfigs;
 import com.ctre.phoenix6.signals.AbsoluteSensorRangeValue;
 // import com.ctre.phoenix.sensors.SensorTimeBase;
 
+import frc.robot.Constants;
 
 public final class CTREConfigs {
   public CANcoderConfiguration swerveCanCoderConfig;
