+++ src/main/java/frc/robot/Robot.java
@@ -9,6 +9,7 @@ import java.io.IOException;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
+import frc.lib.CTREConfigs;
 
 
 /**
